chore: rename package and sort out version numbers (#33)
