+++ packages/extension/package.json
@@ -33,7 +33,7 @@
     "@keplr-wallet/stores": "^0.9.15",
     "@keplr-wallet/types": "^0.9.12",
     "@keplr-wallet/unit": "^0.9.12",
-    "@fetchai/fetchai-wallet-types": "^0.9.10",
+    "@fetchai/wallet-types": "^0.1.0",
     "@fetchai/eth-migration": "^0.8.8",
     "@walletconnect/client": "^1.6.4",
     "aes-js": "^3.1.2",
