+++ packages/provider/package.json
@@ -21,8 +21,8 @@
     "@cosmjs/proto-signing": "^0.24.0-alpha.25",
     "@keplr-wallet/router": "^0.9.6",
     "@keplr-wallet/types": "^0.9.12",
-    "@fetchai/fetchai-wallet-types": "^0.9.10",
-    "@fetchai/umbral-types": "^0.9.10",
+    "@fetchai/wallet-types": "^0.1.0",
+    "@fetchai/umbral-types": "^0.1.0",
     "buffer": "^6.0.3",
     "deepmerge": "^4.2.2",
     "long": "^4.0.0",
