+++ packages/wallet-types/.eslintignore
