+++ packages/wallet-types/.prettierignore
